requirements.txt
black==19.10b0
flake8==3.7.8
flake8-bugbear==19.8.0
Flask==1.1.1
Flask-Limiter==1.1.0
gevent==1.4.0
gnsq==1.0.1
gunicorn==19.9.0
hiredis==1.0.1
ipython==7.8.0
marshmallow==3.2.1
meinheld==1.0.1
mysql-connector-python==8.0.18
psycopg2-binary==2.8.4
pylint==2.4.3
pytest==5.2.1
pytest-cov==2.8.1
python-dateutil==2.8.0
python-json-logger==0.1.11
pytz==2019.3
redis==3.3.10
requests==2.22.0
rq==1.1.0
rq-dashboard==0.5.3
setproctitle==1.1.10
statsd==3.3.0
yoyo-migrations==6.1.0
